FailedToWork/requirements.txt
google-generativeai
duckduckgo-search
rich
